+++ mediation/MyTarget/source/plugin/Assets/GoogleMobileAds/Common/Mediation/myTarget/DummyClient.cs
@@ -1,4 +1,4 @@
-﻿// Copyright 2018 Google LLC
+// Copyright 2018 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -36,7 +36,7 @@ namespace GoogleMobileAds.Common.Mediation.MyTarget
             Debug.Log ("Dummy " + MethodBase.GetCurrentMethod ().Name);
         }
 
-        public bool IsUserConsent()
+        public bool IsConsent()
         {
             Debug.Log ("Dummy " + MethodBase.GetCurrentMethod ().Name);
             return false;