+++ source/plugin/Assets/GoogleMobileAds/Api/AdvertisingBase.cs
@@ -9,7 +9,17 @@ namespace GoogleMobileAds.Api
         public abstract void LoadAd(AdRequest request);
         public abstract bool IsLoaded();
         public abstract void Show();
+        /// <summary>
+        /// Returns the mediation adapter class name.
+        /// </summary>
+        /// <returns></returns>
+        [Obsolete("MediationAdapterClassName() is deprecated, use GetResponseInfo.MediationAdapterClassName() instead.")]
         public abstract string MediationAdapterClassName();
+        /// <summary>
+        /// Returns ad request response info.
+        /// </summary>
+        /// <returns></returns>
+        public abstract ResponseInfo GetResponseInfo();
 
         protected void ExecuteEvent<T>(object sender, EventHandler<T> handler, T args) 
         {