+++ source/plugin/Assets/GoogleMobileAds/Common/IMobileAdsClient.cs
@@ -26,6 +26,9 @@ namespace GoogleMobileAds.Common
         // Initialize the Mobile Ads SDK and mediation adapters.
         void Initialize(Action<IInitializationStatusClient> initCompleteAction);
 
+        // Disable initialization of mediation adapters by the Mobile Ads SDK.
+        void DisableMediationInitialization();
+
         // The application’s audio volume. Affects audio volumes of all ads relative
         // to other audio output. Valid ad volume values range from 0.0 (silent) to 1.0
         // (current device volume). Use this method only if your application has its own