+++ source/plugin/Assets/GoogleMobileAds/Common/MobileAdsEventExecutor.cs
@@ -60,6 +60,7 @@ namespace GoogleMobileAds.Common
                 adEventsQueueEmpty = false;
             }
         }
+
         public void Update()
         {
             if (adEventsQueueEmpty)
@@ -78,7 +79,10 @@ namespace GoogleMobileAds.Common
 
             foreach (Action stagedEvent in stagedAdEventsQueue)
             {
-                stagedEvent.Invoke();
+                if (stagedEvent.Target != null)
+                {
+                    stagedEvent.Invoke();
+                }
             }
         }