+++ source/plugin/Assets/GoogleMobileAds/Platforms/iOS/BannerClient.cs
@@ -28,7 +28,7 @@ namespace GoogleMobileAds.iOS
 
         private IntPtr bannerClientPtr;
 
-#region Banner callback types
+        #region Banner callback types
 
         internal delegate void GADUAdViewDidReceiveAdCallback(IntPtr bannerClient);
 
@@ -44,7 +44,7 @@ namespace GoogleMobileAds.iOS
         internal delegate void GADUAdViewPaidEventCallback(
             IntPtr bannerClient, int precision, long value, string currencyCode);
 
-#endregion
+        #endregion
 
         public event EventHandler<EventArgs> OnAdLoaded;
 
@@ -74,33 +74,34 @@ namespace GoogleMobileAds.iOS
             }
         }
 
-#region IBannerClient implementation
+        #region IBannerClient implementation
 
         // Creates a banner view.
         public void CreateBannerView(string adUnitId, AdSize adSize, AdPosition position)
         {
             this.bannerClientPtr = (IntPtr)GCHandle.Alloc(this);
 
-            switch (adSize.AdType) {
-              case AdSize.Type.SmartBanner:
-                  this.BannerViewPtr = Externs.GADUCreateSmartBannerView(
-                          this.bannerClientPtr, adUnitId, (int)position);
-                  break;
-              case AdSize.Type.AnchoredAdaptive:
-                  this.BannerViewPtr = Externs.GADUCreateAnchoredAdaptiveBannerView(
-                          this.bannerClientPtr,
-                          adUnitId,
-                          adSize.Width,
-                          (int)adSize.Orientation,
-                          (int)position);
-                  break;
-              case AdSize.Type.Standard:
-                  this.BannerViewPtr = Externs.GADUCreateBannerView(
-                          this.bannerClientPtr, adUnitId, adSize.Width, adSize.Height, (int)position);
-                  break;
-              default:
-                  throw new ArgumentException("Invalid AdSize.Type provided.");
-          }
+            switch (adSize.AdType)
+            {
+                case AdSize.Type.SmartBanner:
+                    this.BannerViewPtr = Externs.GADUCreateSmartBannerView(
+                            this.bannerClientPtr, adUnitId, (int)position);
+                    break;
+                case AdSize.Type.AnchoredAdaptive:
+                    this.BannerViewPtr = Externs.GADUCreateAnchoredAdaptiveBannerView(
+                            this.bannerClientPtr,
+                            adUnitId,
+                            adSize.Width,
+                            (int)adSize.Orientation,
+                            (int)position);
+                    break;
+                case AdSize.Type.Standard:
+                    this.BannerViewPtr = Externs.GADUCreateBannerView(
+                            this.bannerClientPtr, adUnitId, adSize.Width, adSize.Height, (int)position);
+                    break;
+                default:
+                    throw new ArgumentException("Invalid AdSize.Type provided.");
+            }
 
             Externs.GADUSetBannerCallbacks(
                     this.BannerViewPtr,
@@ -118,7 +119,8 @@ namespace GoogleMobileAds.iOS
 
             this.bannerClientPtr = (IntPtr)GCHandle.Alloc(this);
 
-            switch (adSize.AdType) {
+            switch (adSize.AdType)
+            {
                 case AdSize.Type.SmartBanner:
                     this.BannerViewPtr = Externs.GADUCreateSmartBannerViewWithCustomPosition(
                     this.bannerClientPtr,
@@ -216,6 +218,11 @@ namespace GoogleMobileAds.iOS
             return Utils.PtrToString(Externs.GADUMediationAdapterClassNameForBannerView(this.BannerViewPtr));
         }
 
+        public IResponseInfoClient GetResponseInfoClient()
+        {
+            return new ResponseInfoClient(this.BannerViewPtr);
+        }
+
         public void Dispose()
         {
             this.DestroyBannerView();
@@ -227,9 +234,9 @@ namespace GoogleMobileAds.iOS
             this.Dispose();
         }
 
-#endregion
+        #endregion
 
-#region Banner callback methods
+        #region Banner callback methods
 
         [MonoPInvokeCallback(typeof(GADUAdViewDidReceiveAdCallback))]
         private static void AdViewDidReceiveAdCallback(IntPtr bannerClient)
@@ -299,7 +306,8 @@ namespace GoogleMobileAds.iOS
                     Value = value,
                     CurrencyCode = currencyCode
                 };
-                AdValueEventArgs args = new AdValueEventArgs() {
+                AdValueEventArgs args = new AdValueEventArgs()
+                {
                     AdValue = adValue
                 };
 
@@ -313,9 +321,7 @@ namespace GoogleMobileAds.iOS
             return handle.Target as BannerClient;
         }
 
-#endregion
+        #endregion
     }
 }
 #endif
-
-