+++ source/plugin/Assets/Plugins/iOS/GADUAdLoader.m
@@ -46,7 +46,7 @@
 - (void)adLoader:(GADAdLoader *)adLoader didFailToReceiveAdWithError:(GADRequestError *)error {
   if (self.adFailedCallback) {
     NSString *errorMsg = [NSString
-        stringWithFormat:@"Failed to receive ad with error: %@", [error localizedFailureReason]];
+        stringWithFormat:@"Failed to receive ad with error: %@", [error localizedDescription]];
     self.adFailedCallback(self.adLoaderClient,
                           [errorMsg cStringUsingEncoding:NSUTF8StringEncoding]);
   }