+++ source/plugin/Assets/Plugins/iOS/GADUInterstitial.m
@@ -62,6 +62,10 @@
   return self.interstitial.responseInfo.adNetworkClassName;
 }
 
+- (GADResponseInfo *)responseInfo {
+  return self.interstitial.responseInfo;
+}
+
 #pragma mark GADInterstitialDelegate implementation
 
 - (void)interstitialDidReceiveAd:(GADInterstitial *)ad {
@@ -72,7 +76,7 @@
 - (void)interstitial:(GADInterstitial *)ad didFailToReceiveAdWithError:(GADRequestError *)error {
   if (self.adFailedCallback) {
     NSString *errorMsg = [NSString
-        stringWithFormat:@"Failed to receive ad with error: %@", [error localizedFailureReason]];
+        stringWithFormat:@"Failed to receive ad with error: %@", [error localizedDescription]];
     self.adFailedCallback(self.interstitialClient,
                           [errorMsg cStringUsingEncoding:NSUTF8StringEncoding]);
   }