+++ source/plugin/Assets/Plugins/iOS/GADURewardedAd.m
@@ -46,9 +46,11 @@
              completionHandler:^(GADRequestError *_Nullable error) {
                if (error) {
                  if (self.adFailedToLoadCallback) {
+                   NSString *errorMsg =
+                       [NSString stringWithFormat:@"Failed to receive ad with error: %@",
+                                                  [error localizedDescription]];
                    self.adFailedToLoadCallback(
-                       self.rewardedAdClient,
-                       [[error localizedFailureReason] cStringUsingEncoding:NSUTF8StringEncoding]);
+                       self.rewardedAdClient, [errorMsg cStringUsingEncoding:NSUTF8StringEncoding]);
                  }
                } else {
                  if (self.adReceivedCallback) {
@@ -75,6 +77,10 @@
   return self.rewardedAd.responseInfo.adNetworkClassName;
 }
 
+- (GADResponseInfo *)responseInfo {
+  return self.rewardedAd.responseInfo;
+}
+
 - (void)rewardedAd:(nonnull GADRewardedAd *)rewardedAd
     userDidEarnReward:(nonnull GADAdReward *)reward {
   if (self.didEarnRewardCallback) {
@@ -89,9 +95,10 @@
 - (void)rewardedAd:(nonnull GADRewardedAd *)rewardedAd
     didFailToPresentWithError:(nonnull NSError *)error {
   if (self.adFailedToShowCallback) {
-    self.adFailedToShowCallback(
-        self.rewardedAdClient,
-        [[error localizedFailureReason] cStringUsingEncoding:NSUTF8StringEncoding]);
+    NSString *errorMsg = [NSString
+        stringWithFormat:@"Failed to present ad with error: %@", [error localizedDescription]];
+    self.adFailedToShowCallback(self.rewardedAdClient,
+                                [errorMsg cStringUsingEncoding:NSUTF8StringEncoding]);
   }
 }