+++ mediation/AdColony/CHANGELOG.md
@@ -1,5 +1,11 @@
 ## AdColony Unity Mediation Plugin Changelog
 
+#### Version 2.0.3
+- Repackaged `adcolony-extras-library.jar` to `adcolony-unity-android-library.aar`, using a `.aar` build to support Unity 2020.
+  * When upgrading to this version, please remove `adcolony-extras-library.jar` from your project.
+- Supports [AdColony Android adapter version 4.1.4.1](https://github.com/googleads/googleads-mobile-android-mediation/blob/master/ThirdPartyAdapters/adcolony/CHANGELOG.md#version-4141).
+- Supports [AdColony iOS adapter version 4.1.5.0](https://github.com/googleads/googleads-mobile-ios-mediation/blob/master/adapters/AdColony/CHANGELOG.md#version-4150).
+
 #### Version 2.0.2
 - Supports [AdColony Android adapter version 4.1.4.1](https://github.com/googleads/googleads-mobile-android-mediation/blob/master/ThirdPartyAdapters/adcolony/CHANGELOG.md#version-4141).
 - Supports [AdColony iOS adapter version 4.1.4.1](https://github.com/googleads/googleads-mobile-ios-mediation/blob/master/adapters/AdColony/CHANGELOG.md#version-4141).