+++ mediation/Chartboost/CHANGELOG.md
@@ -1,5 +1,17 @@
 ## Chartboost Unity Mediation Plugin Changelog
 
+#### Version 3.1.0
+- Supports [Chartboost Android adapter version 8.1.0.0](https://github.com/googleads/googleads-mobile-android-mediation/blob/master/ThirdPartyAdapters/chartboost/CHANGELOG.md#version-8100).
+- Supports [Chartboost iOS adapter version 8.2.0.0](https://github.com/googleads/googleads-mobile-ios-mediation/blob/master/adapters/Chartboost/CHANGELOG.md#version-8200).
+
+#### Version 3.0.1
+- Supports [Chartboost Android adapter version 8.0.3.1](https://github.com/googleads/googleads-mobile-android-mediation/blob/master/ThirdPartyAdapters/chartboost/CHANGELOG.md#version-8031).
+- Supports [Chartboost iOS adapter version 8.1.0.1](https://github.com/googleads/googleads-mobile-ios-mediation/blob/master/adapters/Chartboost/CHANGELOG.md#version-8101).
+
+#### Version 3.0.0
+- Supports [Chartboost Android adapter version 8.0.2.0](https://github.com/googleads/googleads-mobile-android-mediation/blob/master/ThirdPartyAdapters/chartboost/CHANGELOG.md#version-8020).
+- Supports [Chartboost iOS adapter version 8.1.0.0](https://github.com/googleads/googleads-mobile-ios-mediation/blob/master/adapters/Chartboost/CHANGELOG.md#version-8100).
+
 #### Version 2.0.1
 - Supports [Chartboost Android adapter version 7.5.0.1](https://github.com/googleads/googleads-mobile-android-mediation/blob/master/ThirdPartyAdapters/chartboost/CHANGELOG.md#version-7501).
 - Supports [Chartboost iOS adapter version 8.0.4.0](https://github.com/googleads/googleads-mobile-ios-mediation/blob/master/adapters/Chartboost/CHANGELOG.md#version-8040).