+++ mediation/InMobi/CHANGELOG.md
@@ -1,5 +1,21 @@
 ## InMobi Unity Mediation Plugin Changelog
 
+#### Version 3.0.3
+- Supports [InMobi Android adapter version 9.0.7.1](https://github.com/googleads/googleads-mobile-android-mediation/blob/master/ThirdPartyAdapters/inmobi/CHANGELOG.md#version-9071).
+- Supports [InMobi iOS adapter version 9.0.7.2](https://github.com/googleads/googleads-mobile-ios-mediation/blob/master/adapters/InMobi/CHANGELOG.md#version-9072).
+
+#### Version 3.0.2
+- Supports [InMobi Android adapter version 9.0.7.0](https://github.com/googleads/googleads-mobile-android-mediation/blob/master/ThirdPartyAdapters/inmobi/CHANGELOG.md#version-9070).
+- Supports [InMobi iOS adapter version 9.0.7.0](https://github.com/googleads/googleads-mobile-ios-mediation/blob/master/adapters/InMobi/CHANGELOG.md#version-9070).
+
+#### Version 3.0.1
+- Supports [InMobi Android adapter version 9.0.6.0](https://github.com/googleads/googleads-mobile-android-mediation/blob/master/ThirdPartyAdapters/inmobi/CHANGELOG.md#version-9060).
+- Supports [InMobi iOS adapter version 9.0.7.0](https://github.com/googleads/googleads-mobile-ios-mediation/blob/master/adapters/InMobi/CHANGELOG.md#version-9070).
+
+#### Version 3.0.0
+- Supports [InMobi Android adapter version 9.0.4.0](https://github.com/googleads/googleads-mobile-android-mediation/blob/master/ThirdPartyAdapters/inmobi/CHANGELOG.md#version-9040).
+- Supports [InMobi iOS adapter version 9.0.4.0](https://github.com/googleads/googleads-mobile-ios-mediation/blob/master/adapters/InMobi/CHANGELOG.md#version-9040).
+
 #### Version 2.4.0
 - Supports [InMobi Android adapter version 7.3.0.1](https://github.com/googleads/googleads-mobile-android-mediation/blob/master/ThirdPartyAdapters/inmobi/CHANGELOG.md#version-7301).
 - Supports [InMobi iOS adapter version 7.4.0.0](https://github.com/googleads/googleads-mobile-ios-mediation/blob/master/adapters/InMobi/CHANGELOG.md#version-7400).