+++ mediation/Maio/CHANGELOG.md
@@ -1,5 +1,13 @@
 ## maio Unity Mediation Plugin Changelog
 
+#### Version 1.3.4
+- Supports [maio Android adapter version 1.1.11.1](https://github.com/googleads/googleads-mobile-android-mediation/blob/master/ThirdPartyAdapters/maio/CHANGELOG.md#version-11111).
+- Supports [maio iOS adapter version 1.5.4.0](https://github.com/googleads/googleads-mobile-ios-mediation/blob/master/adapters/Maio/CHANGELOG.md#version-1540).
+
+#### Version 1.3.3
+- Supports [maio Android adapter version 1.1.11.1](https://github.com/googleads/googleads-mobile-android-mediation/blob/master/ThirdPartyAdapters/maio/CHANGELOG.md#version-11111).
+- Supports [maio iOS adapter version 1.5.3.1](https://github.com/googleads/googleads-mobile-ios-mediation/blob/master/adapters/Maio/CHANGELOG.md#version-1531).
+
 #### Version 1.3.2
 - Supports [maio Android adapter version 1.1.11.0](https://github.com/googleads/googleads-mobile-android-mediation/blob/master/ThirdPartyAdapters/maio/CHANGELOG.md#version-11110).
 - Supports [maio iOS adapter version 1.5.2.1](https://github.com/googleads/googleads-mobile-ios-mediation/blob/master/adapters/Maio/CHANGELOG.md#version-1521).